# Phase 4.1-4.2: Event Sourcing Foundation & CQRS Implementation