… PHASE_4_1_COMPLETION_REPORT.md
… PHASE_4_2_COMPLETION_REPORT.md
… PHASE_4_IMPLEMENTATION_PLAN.md
… README.md
… infrastructure/database/migrations/003_create_read_models.sql
… infrastructure/kubernetes/deployments/postgresql.yaml
… infrastructure/kubernetes/namespace.yaml
… packages/ai-orchestrator/core/cqrs/commands.rs
… packages/ai-orchestrator/core/cqrs/error.rs
… packages/ai-orchestrator/core/cqrs/handlers.rs
… packages/ai-orchestrator/core/cqrs/mod.rs
… packages/ai-orchestrator/core/cqrs/projections.rs
… packages/ai-orchestrator/core/cqrs/queries.rs
… packages/ai-orchestrator/core/cqrs/read_models.rs
… packages/ai-orchestrator/core/cqrs/tests.rs
… packages/ai-orchestrator/core/event_store/aggregates.rs
… packages/ai-orchestrator/core/event_store/error.rs
… packages/ai-orchestrator/core/event_store/events.rs
… packages/ai-orchestrator/core/event_store/mod.rs
… packages/ai-orchestrator/core/event_store/replay.rs
… packages/ai-orchestrator/core/event_store/serialization.rs
… packages/ai-orchestrator/core/event_store/snapshots.rs
… packages/ai-orchestrator/core/event_store/tests.rs